Divided the setup fill page to api calls, hooks and view
